requirements.txt
langchain>=0.3,<0.4
langgraph>=0.2,<0.3
langchain-groq>=0.2,<0.3
langchain-community>=0.3,<0.4
groq>=0.11
sentence-transformers>=3.0
chromadb>=0.5
pydantic>=2.7
python-dotenv>=1.0
streamlit>=1.36
numpy>=1.26
tabulate>=0.9
